requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
transformers>=4.20.0
torch>=1.12.0
datasets>=2.0.0
flask>=2.0.0
flask-cors>=3.0.0
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.7.0
openai>=1.0.0
